feat: Move to 'pkg' and simplify tests

+++ pkg/lib/bootstrap.sh
@@ -11,16 +11,15 @@ ensure ping google.com -c1 -W2 &>/dev/null
 
 # --------------------- set hostname --------------------- #
 
-read -rp "Choose new hostname: " \
-	-ei "$(</etc/hostname)"
-ensure sudo hostname "$REPLY"
+read -rp "Choose new hostname: " -ei "$(</etc/hostname)"
+ensure sudo hostnamectl set-hostname "$REPLY"
 ensure sudo tee /etc/hostname <<< "$REPLY" &>/dev/null
 
 
 # ----------------------- set hosts ---------------------- #
 
 grep -qe "$REPLY" /etc/hosts || {
-	sudo tee /etc/hosts <<-END >&/dev/null
+	sudo tee -a /etc/hosts <<-EOF >&/dev/null
 	# IP-Address  Full-Qualified-Hostname  Short-Hostname
 	127.0.0.1       localhost
 	::1             localhost ipv6-localhost ipv6-loopback
@@ -29,9 +28,10 @@ grep -qe "$REPLY" /etc/hosts || {
 	ff02::1         ipv6-allnodes
 	ff02::2         ipv6-allrouters
 	ff02::3         ipv6-allhosts
-	END
+	EOF
 
-	read -rp "Check/edit /etc/hosts..." -sn 1
+	echo "Check/edit /etc/hosts..."
+	read -rsn 1
 	sudo "${EDITOR:-vim}" /etc/hosts
 }
 
@@ -39,13 +39,16 @@ grep -qe "$REPLY" /etc/hosts || {
 # --------------------- update fstab --------------------- #
 
 grep -qe '# extra' /etc/fstab || {
-	sudo tee -a /etc/fstab >/dev/null <<-EOF
-	# extra
-	/dev/fox/stg.files  /storage/edwin  xfs  defaults,relatime,X-mount.mkdir=0755  0  2
-	/dev/fox/stg.data  /storage/data  reiserfs defaults,X-mount.mkdir  0 0
-	EOF
+	if [ -e /dev/fox ]; then
+		sudo tee -a /etc/fstab >/dev/null <<-EOF
+		# Extra
+		/dev/fox/stg.files  /storage/edwin  xfs  defaults,relatime,X-mount.mkdir=0755  0  2
+		/dev/fox/stg.data  /storage/data  reiserfs defaults,X-mount.mkdir  0 0
+		EOF
+	fi
 
-	read -rp "Check/edit /etc/fstab..." -sn 1
+	echo "Check/edit /etc/fstab..."
+	read -rsn 1
 	sudo "${EDITOR:-vim}" /etc/fstab
 }
 
@@ -53,21 +56,6 @@ sudo mount -a || {
 	log_error "Error: 'mount -a' failed. Exiting early"
 	[[ -v DEV ]] || exit 1
 }
-
-shopt -q dotglob && was_set=yes
-shopt -u dotglob
-for dir in /storage/edwin/*; do
-	dir="${dir##*/}"
-	ln -sT "/storage/data/$dir" "$HOME/$dir"
-done
-[ "$was_set" = "yes" ] && shopt -s dotglob
-
-ln -sT /storage/data/BraveSoftware "$XDG_CONFIG_HOME/BraveSoftware"
-ln -sT /storage/data/fonts "$XDG_DATA_HOME/fonts"
-ln -sT /storage/data/gnupg "$XDG_DATA_HOME/gnupg"
-ln -sT /storage/data/ssh ~/.ssh
-ln -sT /storage/vault/rodinia/Steam "$XDG_DATA_HOME/Steam"
-ln -sT /storage/data/mozilla/ ~/.mozilla
 
 # ------------------------- date ------------------------- #
 
@@ -85,20 +73,13 @@ ensure sudo locale-gen
 # ------------------------ groups ------------------------ #
 
 # TODO: check /usr/lib/sysgroups.d
-ensure sudo groupadd docker
-ensure sudo groupadd libvirt
-ensure sudo groupadd vboxusers
-ensure sudo groupadd lxd
-ensure sudo groupadd systemd-journal
-ensure sudo groupadd nordvpn
-ensure sudo usermod -aG docker,libvirt,vboxusers,lxd,systemd-journal,nordvpn edwin
-
-# ------------------------ passwd ------------------------ #
-read -rei "${XDG_CONFIG_HOME:-$HOME/.config}"
-read -rp "Choose new hostname: " \
-	-ei "$(</etc/hostname)"
-read- rp "Root Password (enter to skip): "
-[[ -n $REPLY ]] && passwd
+ensure sudo groupadd -f docker
+ensure sudo groupadd -f libvirt
+ensure sudo groupadd -f vboxusers
+ensure sudo groupadd -f lxd
+ensure sudo groupadd -f systemd-journal
+ensure sudo groupadd -f nordvpn
+ensure sudo usermod -aG docker,libvirt,vboxusers,lxd,systemd-journal,nordvpn "$USER"
 
 #
 # ─── INSTALL ACTUAL DOTFILES ────────────────────────────────────────────────────
@@ -106,7 +87,7 @@ read- rp "Root Password (enter to skip): "
 
 # remove anything inherited from /etc/skell
 {
-	mkdir ~/.old
+	mkdir -p ~/.old
 	mv ~/.bash_login ~/.old
 	mv ~/.bash_logout ~/.old
 	mv ~/.bash_profile ~/.old
@@ -114,26 +95,19 @@ read- rp "Root Password (enter to skip): "
 	mv ~/.profile ~/.old
 } >&/dev/null
 
-# temporary git config
-# (in case commits need to be performed)
-<<-EOF cat > ~/.gitconfig
-	[user]
-	    name = Edwin Kofler
-	    email = edwin@kofler.dev
-EOF
-
 [ -d ~/.dots ] || {
 	git clone https://github.com/eankeen/dots ~/.dots
-	cd dots || die "Could not 'cd dots'"
+	cd ~/.dots || die "Could not 'cd dots'"
 	git config --local filter.npmrc-clean.clean "$(pwd)/user/config/npm/npmrc-clean.sh"
+	git config --local filter.slack-term-config-clean.clean "$(pwd)/user/config/slack-term/slack-term-config-clean.sh"
+	git config --local filter.oscrc-clean.clean "$(pwd)/user/config/osc/oscrc-clean.sh"
 }
 
-bm uninstall dotty || log_warn "Could not 'bm uninstall dotty"
-bm install dotty
-dotty reconcile || {
-	die "Error: Could not apply user dotfiles. Exiting"
-}
-
+# TODO: use latest
+curl -LO https://github.com/eankeen/dotty/releases/download/v0.5.0/dotty
+mkdir -p "$XDG_CONFIG_HOME"
+ln -sf ~/.dots/user/config/dotty "$XDG_CONFIG_HOME/dotty"
+./dotty reconcile
 
 #
 # ─── CONCLUSION ─────────────────────────────────────────────────────────────────
@@ -158,7 +132,7 @@ cat <<-EOF
 	  - eankeen/dotty
 
 	Remember:
-	  - \`dots-bootstrap install\`
+	  - dots-bootstrap install
 	  - mkinitcpio
 	  - Initramfs / Kernel (lvm2)
 	  - Root Password
